requirements.txt
asgiref==3.6.0
bokeh==3.1.0
contourpy==1.0.7
Django==4.1.7
django-appconf==1.0.5
django-image-cropping==1.7
django-widget-tweaks==1.4.12
easy-thumbnails==2.8.5
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2023.2
whitenoise==6.4.0
wincertstore==0.2
xyzservices==2023.2.0
